assume we can't use ARMv7 instructions on non-iOS platforms by default
The only Linux/ARM machine I have access to does not support ARMv7.  I
don't know how common that is in general, but this seems like the safe
default.  You can always override it on the command line.

I've also broken the test target into build-test and run-test
subtargets, which can be useful when you're building on a shared
filesystem and running the tests on another machine (without using the
remote-test-host option).  If there's clock skew, the other machine
might try to rebuild stuff unecessarily.  Using run-test avoids
that.